requirements.txt
opencv-python~=4.12.0.88
matplotlib~=3.10.3
Pillow~=11.0.0
numpy~=2.1.2
pandas~=2.3.1
streamlit
transformers
torch~=2.7.1+cu126
torchvision~=0.22.1+cu126
packaging~=25.0
tqdm~=4.67.1
clip~=1.0
ftfy~=6.3.1
regex~=2024.11.6
setuptools~=80.9.0
openai-whisper~=20250625
numba~=0.61.2
more-itertools~=10.7.0
tiktoken~=0.9.0
torchaudio~=2.7.1+cu126
segment_anything~=1.0
fastapi~=0.116.1
uvicorn
python-multipart